requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.3.29
click==8.1.6
clickhouse-connect==0.6.8
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
duckdb==0.8.1
emoji==2.8.0
executing==1.2.0
fastapi==0.85.1
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.40.1
gradio_client==0.4.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langchain==0.0.266
langsmith==0.0.23
linkify-it-py==2.0.2
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.5
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
semantic-version==2.10.0
Send2Trash==1.8.2
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
unstructured==0.10.0
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
